# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zhengji Wellness × Kiri Bar Referral Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /referral/{code}:
    get:
      operationId: getReferralPage
      tags: [referral]
      summary: Get referral landing page data
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Referral page data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralPage"
        "404":
          description: Code not found
...
  /leads/check-duplicate:
    get:
      operationId: checkDuplicate
      tags: [leads]
      summary: Check for duplicate by mobile or membership ID
      parameters:
        - name: mobile
          in: query
          schema:
            type: string
        - name: membershipId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Duplicate check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DuplicateCheck"
...
  /partner/leads:
    get:
      operationId: getPartnerLeads
      tags: [partner]
      summary: Get leads for the authenticated Kiri partner
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Partner's leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
  /partner/commissions:
    get:
      operationId: getPartnerCommissions
      tags: [partner]
      summary: Get commissions for the authenticated Kiri partner
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Partner's commissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionList"
  /partner/stats:
    get:
      operationId: getPartnerStats
      tags: [partner]
      summary: Get partner monthly performance stats
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Partner stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerStats"
  /partner/statement:
    get:
      operationId: getPartnerStatement
      tags: [partner]
      summary: Get partner statement for a month
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Statement data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerStatement"
...
  /partners:
    get:
      operationId: getPartners
      tags: [partner]
      summary: List all Kiri partners (admin/staff)
      responses:
        "200":
          description: Partners list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerProfile"
    post:
      operationId: createPartner
      tags: [partner]
      summary: Create a Kiri partner account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Partner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerProfile"
  /partners/{id}:
    get:
      operationId: getPartner
      tags: [partner]
      summary: Get partner profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Partner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerProfile"
...
components:
  schemas:
...
    ReferralPage:
      type: object
      required: [partnerName, referralCode, offer, campaignActive]
      properties:
        partnerName:
          type: string
        kirimembershipId:
          type: string
        referralCode:
          type: string
        offer:
          type: string
        offerZh:
          type: string
        discountPercent:
          type: number
        campaignActive:
          type: boolean
        campaignEndDate:
          type: ["string", "null"]
...
    Lead:
      type: object
      required: [id, name, mobile, stage, referralCode, partnerId, consentGiven, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameZh:
          type: ["string", "null"]
        mobile:
          type: string
        whatsapp:
          type: ["string", "null"]
        kirimembershipId:
          type: ["string", "null"]
        referralCode:
          type: string
        partnerId:
          type: integer
        partnerName:
          type: string
        stage:
          type: string
          enum: [new_lead, appointment_booked, arrived, free_consultation_only, first_paid_treatment, package_purchased, invalid_cancelled]
        selectedOffer:
          type: string
        appointmentIntent:
          type: ["string", "null"]
        appointmentDate:
          type: ["string", "null"]
        netSaleAmount:
          type: ["number", "null"]
        proofUrl:
          type: ["string", "null"]
        consentGiven:
          type: boolean
        notes:
          type: ["string", "null"]
        lang:
          type: string
        commissionId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadList:
      type: object
      required: [leads, total, page, limit]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DuplicateCheck:
      type: object
      required: [isDuplicate]
      properties:
        isDuplicate:
          type: boolean
        field:
          type: ["string", "null"]
        existingLeadId:
          type: ["integer", "null"]
    Commission:
      type: object
      required: [id, leadId, partnerId, amount, commissionType, status, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        leadName:
          type: string
        partnerId:
          type: integer
        partnerName:
          type: string
        amount:
          type: number
        commissionType:
          type: string
          enum: [flat_rm30, package_percent]
        commissionRate:
          type: ["number", "null"]
        netSaleAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, paid, disputed, rejected]
        payoutReference:
          type: ["string", "null"]
        proofUrl:
          type: ["string", "null"]
        auditNote:
          type: ["string", "null"]
        approvedBy:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CommissionList:
      type: object
      required: [commissions, total, page, limit]
      properties:
        commissions:
          type: array
          items:
            $ref: "#/components/schemas/Commission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    PartnerStats:
      type: object
      required: [month, totalLeads, totalArrivals, totalConversions, totalCommissionPending, totalCommissionApproved, totalCommissionPaid, conversionRate]
      properties:
        month:
          type: string
        totalLeads:
          type: integer
        totalArrivals:
          type: integer
        totalConversions:
          type: integer
        totalCommissionPending:
          type: number
        totalCommissionApproved:
          type: number
        totalCommissionPaid:
          type: number
        conversionRate:
          type: number
        tierBonus:
          type: ["number", "null"]
        tierBonusLabel:
          type: ["string", "null"]
    PartnerStatement:
      type: object
      required: [month, partnerName, referralCode, leads, commissions, summary]
      properties:
        month:
          type: string
        partnerName:
          type: string
        referralCode:
          type: string
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        commissions:
          type: array
          items:
            $ref: "#/components/schemas/Commission"
        summary:
          $ref: "#/components/schemas/PartnerStats"
...
    PartnerProfile:
      type: object
      required: [id, userId, displayName, referralCode, kirimembershipId, isActive, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        displayName:
          type: string
        referralCode:
          type: string
        kirimembershipId:
          type: string
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        totalLeads:
          type: integer
        totalConversions:
          type: integer
        totalCommissionEarned:
          type: number
        createdAt:
          type: string
    PartnerInput:
      type: object
      required: [displayName, kirimembershipId, username, password]
      properties:
        displayName:
          type: string
        kirimembershipId:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
...