# lib/api-zod/src/generated/types/duplicateCheck.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Zhengji Wellness × Kiri Bar Referral Management API
 * OpenAPI spec version: 0.1.0
 */

export interface DuplicateCheck {
  isDuplicate: boolean;
  /** @nullable */
  field?: string | null;
  /** @nullable */
  existingLeadId?: number | null;
}