# lib/api-zod/src/generated/types/referralPage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Zhengji Wellness × Kiri Bar Referral Management API
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralPage {
  partnerName: string;
  kirimembershipId?: string;
  referralCode: string;
  offer: string;
  offerZh?: string;
  discountPercent?: number;
  campaignActive: boolean;
  /** @nullable */
  campaignEndDate?: string | null;
}